rename layout -> styles and fix imports
