+++ src/components/RegisterRestaurant.jsx
@@ -10,7 +10,7 @@ import apiUrl from '../utils/apiUrl';
 import { restaurantValidationSchema }
     from '../utils/formValidationSchemas';
 
-import createStyles from '../styles/layout';
+import createStyles from '../styles/styles';
 import { Button } from './ui/Buttons';
 import { Input, PasswordInput } from './ui/InputFields';
